+++ test/cwl_workflows/test_cat/test_cat.cwl
@@ -1,0 +1,21 @@
+cwlVersion: v1.2
+class: CommandLineTool
+
+hints:
+  DockerRequirement:
+    dockerPull: ubuntu:18.04
+  InlineJavascriptRequirement: {}
+
+
+inputs:
+  file:
+    type: File
+
+baseCommand: [cat, file]
+
+outputs:
+  an_output_name:
+    type: stdout
+
+stdout: a_stdout_file
+
