+++ test/cwl_workflows/test_single_step/test_single_step.cwl
@@ -1,11 +1,11 @@
-cwlVersion: v1.1
+cwlVersion: v1.2
 class: CommandLineTool
 
 hints:
   DockerRequirement:
     dockerPull: ubuntu:18.04
+  InlineJavascriptRequirement: {}
 
-baseCommand: [touch]
 
 inputs:
   name:
@@ -13,8 +13,14 @@ inputs:
     inputBinding:
       position: 0
 
+baseCommand: [touch]
+
+arguments:
+  - valueFrom: $(inputs.name).txt
+
 outputs:
   testing_result:
     type: File
     outputBinding:
-      glob: $(inputs.name)
+      glob: "*.txt"
+
